[refactoring] Add strategy pattern
