+++ SSD_Driver/buffer_manager.h
@@ -1,0 +1,32 @@
+#pragma once
+#include <string>
+#include <vector>
+
+struct Command {
+	std::string type;
+	bool isEmpty = true;
+	int address = -1;
+	unsigned int value = 0;
+	int eraseSize = 0;
+};
+
+class BufferManager {
+public:
+	BufferManager(std::string nandPath);
+
+	void initializeBuffer();
+	void loadBufferData();
+	void parsingCommandBuffer(const std::vector<std::string> parts);
+	void updateBufferFile();
+	void addCommandInBuffer(const std::string& command_type, int address, int valueOrSize);
+	void updateBufferWithNewCommands(const std::vector<Command>& commands_to_add);
+	bool readFromBuffer(int address, unsigned int& outValue);
+	void flush();
+
+private:
+	const std::string bufferDir = "buffer";
+	std::vector<Command> commands;
+	std::string nandPath;
+	const int MAX_ERASE_SIZE = 10;
+	const int MAX_BUFFER_SIZE = 5;
+};
